renaming sort dialog to main menu
