+++ lib/uiComponents/ProgramGuideNative/actions/handleEsc.js
@@ -1,8 +1,8 @@
 export default ({ component }) => event => {
-  const { toggleSortDialog, closeEventDialog } = component
+  const { toggleMainMenu, closeEventDialog } = component
   if (event.key === 'Esc' || event.key === 'Escape') {
     event.preventDefault()
-    toggleSortDialog(false)
+    toggleMainMenu(false)
     closeEventDialog()
   }
 }
