+++ lib/uiComponents/ProgramGuideNative/actions/index.js
@@ -7,7 +7,7 @@ export { default as getHandleHashChange } from './handleHashChange'
 export { default as getShowFloorPlan } from './showFloorPlan'
 export { default as getToggleFloorPlanInteractive } from './toggleFloorPlanInteractive'
 export { default as getToggleHeaders } from './toggleHeaders'
-export { default as getToggleSortDialog } from './toggleSortDialog'
+export { default as gettoggleMainMenu } from './toggleMainMenu'
 export { default as getOpenEventDialog } from './openEventDialog'
 export { default as getCloseEventDialog } from './closeEventDialog'
 export { default as getSelectActiveLocation } from './selectActiveLocation'
