+++ lib/uiComponents/ProgramGuideNative/actions/selectActiveLocation.js
@@ -1,6 +1,6 @@
 export default ({ setState, component }) => async _location => {
-  const { toggleSortDialog } = component
-  toggleSortDialog(false) // close upon selection
+  const { toggleMainMenu } = component
+  toggleMainMenu(false) // close upon selection
   if (state.sortBy === 'location' && state.view === 'guide') return
   await setState({
     view: 'guide',
