+++ lib/uiComponents/ProgramGuideNative/actions/selectSortBy.js
@@ -1,10 +1,10 @@
 import { getToday } from '../../../utilities'
 
 export default ({ state, setState, component }) => async sortBy => {
-  const { toggleSortDialog } = component
+  const { toggleMainMenu } = component
 
   // close upon selection
-  toggleSortDialog(false)
+  toggleMainMenu(false)
   if (state.sortBy === sortBy && state.view === 'guide') return
   const currentDay = sortBy === 'now' || sortBy === 'nowOnly' ? getToday(state) : state.activeDay
   await setState({
