+++ lib/uiComponents/ProgramGuideNative/actions/toggleMainMenu.js
@@ -1,11 +1,11 @@
 import { clearHTML, parseHTML } from '../../../utilities'
-import { renderSortDialog } from '../render'
+import { renderMainMenu } from '../render'
 
 export default ({ state, component }) => _bool => {
   const { root } = component
   const bool = _bool === 'true' ? true : !!_bool
   const dialog = root.querySelector('.pg-dialog--sort')
-  const dialogContent = parseHTML(renderSortDialog(state, component))
+  const dialogContent = parseHTML(renderMainMenu(state, component))
   clearHTML(dialog)
   dialog.appendChild(dialogContent)
   if (bool) dialog.showModal()
