+++ lib/uiComponents/ProgramGuideNative/index.js
@@ -13,7 +13,7 @@ import {
   getShowFloorPlan,
   getOpenEventDialog,
   getToggleFloorPlanInteractive,
-  getToggleSortDialog,
+  gettoggleMainMenu,
   getCloseEventDialog,
   getSelectActiveLocation,
   getSelectActiveDay,
@@ -82,7 +82,7 @@ class ProgramGuideNative extends HTMLElement {
       showFloorPlan: getShowFloorPlan,
       toggleFloorPlanInteractive: getToggleFloorPlanInteractive,
       toggleHeaders: getToggleHeaders,
-      toggleSortDialog: getToggleSortDialog,
+      toggleMainMenu: gettoggleMainMenu,
       openEventDialog: getOpenEventDialog,
       closeEventDialog: getCloseEventDialog,
       selectActiveLocation: getSelectActiveLocation,
