+++ lib/uiComponents/ProgramGuideNative/render/index.js
@@ -1,6 +1,6 @@
 import { DEFAULT_BREAKPOINT } from '../../../constants'
 import { getStyles } from '../getStyles'
-import { default as renderSortDialog } from './renderSortDialog'
+import { default as renderMainMenu } from './renderMainMenu'
 import { default as renderEventDialog } from './renderEventDialog'
 import { default as renderByLocation } from './renderByLocation'
 import { default as renderByTime } from './renderByTime'
@@ -60,7 +60,7 @@ const getRender = ({ state, component }) => async ({ initializing } = {}) => {
     </style>
     <section id="ProgramGuide" class="pg-body ${state.maximized ? 'pg-maximized' : ''}">
       <header class="pg-header">
-        <button class="pg-bars-menu-button" onclick="${runMethod('toggleSortDialog', true)}">
+        <button class="pg-bars-menu-button" onclick="${runMethod('toggleMainMenu', true)}">
           <i class="pg-more-icon" title="Main Menu"></i>
         </button>
         ${renderSearchBar(state)}
@@ -76,7 +76,7 @@ const getRender = ({ state, component }) => async ({ initializing } = {}) => {
       }
       ${renderMenu(state)}
     </section>
-    <dialog class="pg-dialog pg-dialog--sort">${renderSortDialog(state, component)}</dialog>
+    <dialog class="pg-dialog pg-dialog--sort">${renderMainMenu(state, component)}</dialog>
     <dialog class="pg-dialog pg-dialog--event">${renderEventDialog(state)}</dialog>
   `
   clearHTML(root) // clear loading spinner
@@ -122,7 +122,7 @@ const getRender = ({ state, component }) => async ({ initializing } = {}) => {
 }
 
 export {
-  renderSortDialog,
+  renderMainMenu,
   renderEventDialog,
   getRender,
 }
