+++ lib/uiComponents/ProgramGuideNative/render/renderMainMenu.js
@@ -1,7 +1,7 @@
 import { DEFAULT_TIMES } from '../../../constants'
 import { mapString, runMethod, parseLocationHash } from '../../../utilities'
 
-const renderSortDialog = ({ locations, hiddenLocations, times, activeDay, activeTime: _activeTime, sortBy }, component) => {
+const renderMainMenu = ({ locations, hiddenLocations, times, activeDay, activeTime: _activeTime, sortBy }, component) => {
 
   // only show location list if the columns are collapsed
   const { days, interval } = times || DEFAULT_TIMES
@@ -29,7 +29,7 @@ const renderSortDialog = ({ locations, hiddenLocations, times, activeDay, active
     <section class="pg-dialog-body">
       <header class="pg-dialog-header">
         <h1 class="pg-dialog-title">Options</h1>
-        <button class="pg-dialog-close" title="Close" onclick="${runMethod('toggleSortDialog', false)}"><i class="pg-close-icon" title="Close"></i></button>
+        <button class="pg-dialog-close" title="Close" onclick="${runMethod('toggleMainMenu', false)}"><i class="pg-close-icon" title="Close"></i></button>
       </header>
       <section class="pg-dialog-content">
         <menu class="pg-dialog-menu">
@@ -99,4 +99,4 @@ const renderSortDialog = ({ locations, hiddenLocations, times, activeDay, active
   `
 }
 
-export default renderSortDialog
+export default renderMainMenu
